storybookGenerator/requirements.txt
fastapi
uvicorn
openai
pillow
python-dotenv
requests
